Move stat var hierarchy componenets into a separate folder
