+++ static/js/browser/types.ts
@@ -40,39 +40,6 @@ export interface SourceSeries {
   unit?: string;
 }
 
-export interface StatVarInfo {
-  id: string;
-  specializedEntity: string;
-  displayName: string;
-}
-
-export interface StatVarGroupInfo {
-  id: string;
-  specializedEntity: string;
-  displayName: string;
-}
-
-export interface StatVarGroupNodeType {
-  absoluteName: string;
-  level: number;
-  parent?: string;
-  childStatVarGroups?: Array<{ id: string; specializedEntity: string }>;
-  childStatVars?: StatVarNodeType[];
-}
-
-export interface StatVarNodeType {
-  displayName: string;
-  id: string;
-  searchName: string;
-  parent: string;
-  level: number;
-}
-
-export enum StatVarHierarchyNodeType {
-  STAT_VAR_GROUP,
-  STAT_VAR,
-}
-
 export enum PageDisplayType {
   PLACE_STAT_VAR,
   PLACE_WITH_WEATHER_INFO,
