+++ static/js/stat_var_hierarchy/stat_var_search.tsx
